requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djoser
django-cors-headers
django-cleanup
django-concurrency
drf-nested-routers